requirements.txt
astroid==2.12.14
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
docopt-ng==0.8.1
easywatch==0.0.5
Flask==2.2.2
Flask-Minify==0.41
Flask-WTF==1.0.1
htmlmin==0.1.12
idna==3.4
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
lazy-object-proxy==1.9.0
lesscpy==0.15.1
MarkupSafe==2.1.1
mccabe==0.7.0
mysql-connector-python==8.0.31
numpy==1.24.1
pandas==1.5.2
platformdirs==2.6.2
ply==3.11
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2022.7
rcssmin==1.1.1
requests==2.28.1
six==1.16.0
tomlkit==0.11.6
urllib3==1.26.13
watchdog==2.2.0
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
xxhash==3.1.0
